light: factor out syslog_ng_local_executor
